move `CommentBase`
